requirements.txt
streamlit>=1.42.0
pandas>=2.2.0
pydantic>=2.10.0
python-dotenv>=1.0.1

# Optional parsing helpers
unstructured[pdf]>=0.16.11
pypdf>=5.4.0
pdfplumber>=0.11.0
pytesseract>=0.3.13
Pillow>=10.4.0
